splitWiseBackendNew/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0 
passlib[bcrypt]==1.7.4        
redis==5.0.4
email-validator==2.1.1